+++ src/components/cards/AssetCard.tsx
@@ -9,6 +9,7 @@ import { Badge, Card, Group } from '@mantine/core'
 import moment from 'moment'
 
 import { useCurrencyValue } from 'src/hooks/useCurrencyValue'
+import { useFullyRentedAPR } from 'src/hooks/useFullyRentedAPR'
 import { selectUserRentCalculation } from 'src/store/features/settings/settingsSelector'
 import { UserRealtoken } from 'src/store/features/wallets/walletsSelector'
 import { RentCalculationState } from 'src/types/RentCalculation'
@@ -48,8 +49,7 @@ const AssetCardComponent: FC<AssetCardProps> = (props) => {
   const yearlyAmount = props.value.amount * props.value.netRentYearPerToken
   const totalInvestment = props.value.totalInvestment
 
-  const fullyRentedAPR =
-    Math.floor(fullyRentedAPREstimation(props.value) * 100) / 100
+  const fullyRentedAPR = useFullyRentedAPR(props.value)
 
   return (
     <Card
@@ -155,7 +155,9 @@ const AssetCardComponent: FC<AssetCardProps> = (props) => {
 
       <div className={styles.groupApart}>
         <div className={styles.textSm}>{t('fullyRentedEstimation')}*</div>
-        <div className={styles.textSm}>{fullyRentedAPR} %</div>
+        <div className={styles.textSm}>
+          {tNumbers('percent', { value: fullyRentedAPR })}
+        </div>
       </div>
 
       <div style={{ flex: '1 1 auto' }} />
@@ -164,50 +166,6 @@ const AssetCardComponent: FC<AssetCardProps> = (props) => {
       <div className={styles.textLocation}>{props.value.fullName}</div>
     </Card>
   )
-}
-
-const fullyRentedAPREstimation = (token: UserRealtoken) => {
-  if (token.rentedUnits === token.totalUnits) {
-    return token.annualPercentageYield
-  }
-
-  if (token.rentedUnits === 0 && token.annualPercentageYield !== 0) {
-    return token.annualPercentageYield
-  }
-
-  const APREstimation = () => {
-    if (token.history.length > 0) {
-      let propInfo = token.history[0].values
-      const history = token.history.map((h) => {
-        propInfo = { ...propInfo, ...h.values }
-        return propInfo
-      })
-
-      const previousAPR = history
-        .map((h) => {
-          if (
-            h.rentedUnits &&
-            h.rentedUnits !== 0 &&
-            h.netRentYear &&
-            h.tokenPrice
-          ) {
-            return (
-              ((h.netRentYear * token.totalUnits) /
-                (token.totalTokens * h.tokenPrice * h.rentedUnits)) *
-              100
-            )
-          }
-          return 0
-        })
-        .filter((apr) => apr !== undefined)
-
-      return Math.max(...previousAPR)
-    } else {
-      return 0
-    }
-  }
-
-  return APREstimation()
 }
 
 export const AssetCard = memo(AssetCardComponent)
